+++ Src/Private/Get-AbrOntapNetworkBdomain.ps1
@@ -5,7 +5,7 @@ function Get-AbrOntapNetworkBdomain {
     .DESCRIPTION
 
     .NOTES
-        Version:        0.6.7
+        Version:        0.6.12
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -38,6 +38,10 @@ function Get-AbrOntapNetworkBdomain {
                     $BDomainObj += [pscustomobject]$inobj
                 }
 
+                if ($Healthcheck.Network.Port) {
+                    $BDomainObj | Where-Object { $null -eq $_.'Failover Groups' -and $null -eq $_.'Ports' } | Set-Style -Style Warning
+                }
+
                 $TableParams = @{
                     Name = "Network Broadcast Domain - $($ClusterInfo.ClusterName)"
                     List = $false
@@ -47,6 +51,15 @@ function Get-AbrOntapNetworkBdomain {
                     $TableParams['Caption'] = "- $($TableParams.Name)"
                 }
                 $BDomainObj | Table @TableParams
+                if ($Healthcheck.Network.Port -and ($BDomainObj | Where-Object { $null -eq $_.'Failover Groups' -and $null -eq $_.'Ports' })) {
+                    Paragraph "Health Check:" -Bold -Underline
+                    BlankLine
+                    Paragraph {
+                        Text "Best Practice:" -Bold
+                        Text " Broadcast Domains should have associated Failover Groups and Ports assigned to them, review the highlighted Broadcast Domains above and take corrective action as necessary."
+                    }
+                    BlankLine
+                }
             }
         } catch {
             Write-PScriboMessage -IsWarning $_.Exception.Message